+++ src/main/java/chess/controller/ChessController.java
@@ -2,7 +2,7 @@ package chess.controller;
 
 import chess.model.board.ChessBoard;
 import chess.model.board.ChessBoardInitializer;
-import chess.model.piece.Points;
+import chess.model.board.Points;
 import chess.model.position.ChessPosition;
 import chess.view.GameArguments;
 import chess.view.GameCommand;
